fixes to mute function

+++ commands/moderation/mute.js
@@ -1,23 +1,27 @@
 const timeout = ms => new Promise(resolve => setTimeout(resolve, ms));
 
 const mute = async (client, message, args) => {
-  const { deletable, guild, mentions, channel } = message;
+  const { deletable, guild, mentions, channel, author } = message;
+
+  if (author.id !== guild.owner.id)
+    return message.reply("you don't have permissions to mute members.");
+
   if (deletable) message.delete();
 
   const toMute = mentions.members.first();
   if (!toMute)
     return message.reply("you must specify a member").then(m => m.delete(5000));
 
-  let mutedRole = guild.roles.find(r => r.name === "muted");
+  let mutedRole = guild.roles.find(r => r.name === "MUTED");
   if (!mutedRole) {
     try {
       mutedRole = await guild.createRole({
-        name: "muted",
+        name: "MUTED",
         permissions: [],
         position: 999
       });
       message.guild.channels.forEach(async (channel, id) => {
-        await channel.overwritePermissions(mutedRole, {
+        const test = await channel.overwritePermissions(mutedRole, {
           SEND_MESSAGES: false,
           ADD_REACTIONS: false,
           SEND_TTS_MESSAGES: false,
@@ -36,14 +40,14 @@ const mute = async (client, message, args) => {
   try {
     // Add the role to the user
     await toMute.addRole(mutedRole);
-    channel.send(
-      `${toMute}, you've been muted for 5minutes... go and sit in the corner.`
-    );
+    channel.send(`${toMute}, you've been muted for 5 minutes...`);
 
     // Remove the role after 5000
     await timeout(300000);
     await toMute.removeRole(mutedRole);
-    channel.send(`Alright ${toMute} - I hope you've learnt your lesson.`);
+    channel.send(
+      `Alright ${toMute}, you've been unmuted... I hope you've learnt your lesson.`
+    );
   } catch (error) {
     console.error(error);
   }
